Magic Trick! We didn't like the previous option. Too complicated
Turns out, we can make a submodule in RecordChain, parametrized by a
predicate that tells us whether a record belongs in the current pool
of records. Then, we define a recordstore separately, and prove lemmas
s2 and s3 for a SPECIFIC pool of records.

Next step is to take the postulates out into proper datatypes and make
them module parameters.