Started working on the concrete model; still not complete. The places where records must be abstracted are still a bit misterious to me.
For example; we could define the Concrete.Extends relation
over concrete or abstract records;

It does seem that there is
a slight break of symmetry, however: In Abstract; we care only about
blocks, qcs and the initial record. In conrete, records can be more
things.
It does seem to me that we should make "storeable concrete records",
which can only be blocks and qcs and are meant to be stored in the recordstore.