Progress on relating concrete BlockTree to abstract (notes below)
IsValidQC is a function of an EpochConfig.  Therefore, to keep IsValidQC proofs as metadata with a
BlockTree, the BlockTree must know its EpochConfig.  I have added a Meta EpochConfig to BlockTree
for this purpose.  Notes:

- Ultimately, the abstract EpochConfig will be a function of a peer's state (EventProcessor), but
  I'm not sure which parts of it, nor whether we want to pass it around everywhere, so I settled on
  making it explicitly metadata in the BlockTree.
- I added this Meta EpochConfig ant the end of the BoockTree record and also moved btIdToQuorumCert
  after it.  This is because it seems that the fields for which we create lenses must be a prefix of
  the fields in the record (I get errors otherwise), and because we don't want to keep proofs in
  fields with lenses.  Ideally we could avoid this to keep the model and implementation
  closer (while keeping both close to the Rust code).  But I don't anticipate many instances of this
  problem, and it's not a huge one, so no problem if the is more hassle than it's worth.
- It might be preferable to add the EpochConfig as a parameter of the BlockTree record, but I could
  not make this work with the lens definitions.

I had some trouble with cyclic imports, so I sucked the definition of IsValidQC and friends into
Consensus.Types.  Not ideal but not sure how to avoid it.

I've backed off the use of lenses in some places as they complicate the proofs (too much for me to
readily understand how to deal with them anyway).  The most important use of lenses is in the actual
algorithm model, to keep it as close to the Haskell code as we can, so it's not a big deal to not
use them in proofs, I think.

The notation I introduced for "downgrading" a lens to an accessor function is ugly and annoying.
First, I haven't done anything with operator precedence, so I need parens every time I downgrade a
lens.  Second, the specific character I happened upon and used (⇣), while intuitively nice, is
annoying because I seem to get a variety of very similar-looking --- but different -- characters by
tying \d in Agda.  So I keep getting out-of-scope errors for things that look fine.  Victor, I'm
sure you can make this more beautiful and less annoying :-).