Made meta level-polymorphic; added a bunch of comments.
I'm convinced carrying the epoch config as a parameter
is a bad idea. I understand we want the types to be as close
as possible to Haskell, but if we want to carry proofs about
the validity of the objects inside the blocktree; the blocktree must
receive the epoch config as a parameter.