Propagating changes and trying to rework the proofs.
We actually have two insertion functions: one that inserts
a valid qc and one that inserts a linkableBlock.

We need to prove that both respect the invariants. Note that the
insertion function must work over /concrete/ records if we are to
use it later on. I'm not confident this is the best way to
join the pieces yet though, i will think about it