Isolate and demonstrate Agda bug
On the previous commit, if we try to build Everyting.agda, it fails with the following message:

  An internal error has occurred. Please report this as a bug.
  Location of the error: src/full/Agda/TypeChecking/Serialise/Instances/Internal.hs:148

It doesn't even tell us which file caused the problem, so we had to bisect to find that it was
ModelDraft.agda.

In this commit, we comment out some weird unfinished stuff that eliminates the problem.  We could
consider trying to reproduce this with a small example and report it.