Banish EpochConfig from Yasm; details below
EpochConfigs are a purely application-specific notion.  Here, we remove all traces of EpochConfig
from Yasm.  Therefore, AvailableEpochs and the related e parameter for SystemState are gone.
Similary, there is no longer a step-epoch.

To enable proving properties about messages signed for honest public keys, the application provides
a type ValidSenderForPK (and a proof that it is stable), which captures which peers can sign which
messages for which public keys in a given SystemState.

Such properties are in an inner module in Yasm.Properties that receives a
StepPeerState-AllValidParts as a module parameter.  This provides evidence that, if a step of the
implementation introduces a new signature, then ValidSenderForPK holds for the *poststate* of the
step, allowing for the possibility that, in a single step, a peer learns that it is allowed to sign
certain messages for a particular PK (in our case, this could be because it commits an
epoch-changing transaction and creates an EpochChangeProof, thus allowing it to sign messages for PK
in a new epoch) and signs and sends a message.

Existing properties are updated and fixed, and a few new ones are introduced.

Signed-off-by: Mark Moir <mark.moir@oracle.com>