Make Abstract module provide evidence of specific injectivity failure (#166)
This pull request makes the Abstract properties hold unless they can provide a specific example of noninjectivity of abstract Blocks that are "in the system", according to a predicate provided as a module parameter. This removes the possibility for the proofs to cheat by simply creating a pair of different Blocks with the same bId, which is trivial without the requirement to show that those Blocks are considered to be "in the system" by whomever is instantiating the Abstract properties. The top-level Abstract modules now take a module parameter that asserts that there are no injectivity failures among Records that are "in the system", which allows us to eliminate the possibility of such an injectivity failure., yielding the desired correctness properties. We have yet to push this up to the implementation we are modelling, so this gap is "papered over" for now using obm-damgerous-magic.